+++ CHANGELOG.rst
@@ -14,6 +14,15 @@ Change Log
 Unreleased
 **********
 
+[1.5.0] - 2022-10-26
+********************
+
+Changed
+=======
+* Remove redundant lookup of signal in consumer loop (should not have any effect)
+* Explicitly encode message header values as UTF-8 (no change in behavior)
+* Log full event data on all producer errors
+
 [1.4.3] - 2022-10-31
 ********************
 
@@ -41,6 +50,7 @@ Fixed
 
 Changed
 =======
+
 * Remove override of auto.offset.reset on consumer (which will default to "latest"). New consumer groups will consume only messages that are sent after the group was initialized.
 * Remove redundant lookup of signal in consumer loop (should not have any effect)
 * Explicitly encode message header values as UTF-8 (no change in behavior)